chore: uriBuilder 파일 폴더 변경 반영
